Expose port 4180 inside the docker container
Ansible does a strange thing with the `publish` parameter. If a port
inside the docker container has not been explicitly `EXPOSE`d, the
`publish` parameter will ignore this port.

This contradicts the expection set by the docker cli.

The docker cli, in these circumstances, expose the port since this is
implied by the caller.